+++ ttlser/ttlser/__init__.py
@@ -1,2 +1,2 @@
-from .ttlser import *
-from .ttlser import __all__
+from .serializers import *
+from .serializers import __all__
